requirements.txt
aniso8601==8.0.0
astroid==2.4.1
attrs==19.3.0
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
flasgger==0.9.4
Flask==1.1.2
Flask-RESTful==0.3.8
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.3.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pybase62==0.4.3
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
flake8
black
